+++ test/component.test.js
@@ -19,20 +19,20 @@ describe('Father.ComponentPackage', function() {
     pkg.files['src/test.js'].dependencies.should.eql(['to-function']);
     var pkgDeps = pkg.dependencies;
     Object.keys(pkgDeps).should.eql(['each', 'to-function']);
-    pkgDeps['each'].should.eql(pkg.get('each@0.2.3'));
-    pkgDeps['to-function'].should.eql(pkg.get('to-function@1.0.0'));
+    pkgDeps['each'].should.eql(pkg.getPackage('each@0.2.3'));
+    pkgDeps['to-function'].should.eql(pkg.getPackage('to-function@1.0.0'));
 
-    var each = pkg.get('each@0.2.3');
+    var each = pkg.getPackage('each@0.2.3');
     each.main.should.eql('index.js');
     each.name.should.eql('each');
     each.version.should.eql('0.2.3');
     each.files['index.js'].dependencies.should.eql(['type', 'to-function']);
     var eachDeps = each.dependencies;
     Object.keys(eachDeps).should.eql(['to-function', 'type']);
-    eachDeps['type'].should.eql(pkg.get('type@1.0.0'));
-    eachDeps['to-function'].should.eql(pkg.get('to-function@2.0.0'));
+    eachDeps['type'].should.eql(pkg.getPackage('type@1.0.0'));
+    eachDeps['to-function'].should.eql(pkg.getPackage('to-function@2.0.0'));
 
-    var type = pkg.get('type@1.0.0');
+    var type = pkg.getPackage('type@1.0.0');
     type.main.should.eql('index.js');
     type.name.should.eql('type');
     type.version.should.eql('1.0.0');
@@ -40,7 +40,7 @@ describe('Father.ComponentPackage', function() {
     var typeDeps = type.dependencies;
     Object.keys(typeDeps).should.eql([]);
 
-    var props = pkg.get('props@1.1.2');
+    var props = pkg.getPackage('props@1.1.2');
     props.main.should.eql('index.js');
     props.name.should.eql('props');
     props.version.should.eql('1.1.2');
@@ -48,7 +48,7 @@ describe('Father.ComponentPackage', function() {
     var propsDeps = props.dependencies;
     Object.keys(propsDeps).should.eql([]);
 
-    var toFunction = pkg.get('to-function@1.0.0');
+    var toFunction = pkg.getPackage('to-function@1.0.0');
     toFunction.main.should.eql('index.js');
     toFunction.name.should.eql('to-function');
     toFunction.version.should.eql('1.0.0');
@@ -56,14 +56,14 @@ describe('Father.ComponentPackage', function() {
     var toFunctionDeps = toFunction.dependencies;
     Object.keys(toFunctionDeps).should.eql([]);
 
-    var toFunction2 = pkg.get('to-function@2.0.0');
+    var toFunction2 = pkg.getPackage('to-function@2.0.0');
     toFunction2.main.should.eql('index.js');
     toFunction2.name.should.eql('to-function');
     toFunction2.version.should.eql('2.0.0');
     toFunction2.files['index.js'].dependencies.should.eql(['props']);
     var toFunction2Deps = toFunction2.dependencies;
     Object.keys(toFunction2Deps).should.eql(['props']);
-    toFunction2Deps['props'].should.eql(pkg.get('props@1.1.2'));
+    toFunction2Deps['props'].should.eql(pkg.getPackage('props@1.1.2'));
   });
 
   it('ref version', function() {