+++ test/spm.test.js
@@ -19,44 +19,44 @@ describe('Father.SpmPackage', function() {
     pkg.version.should.eql('1.0.0');
     pkg.files['a.js'].dependencies.should.eql(['b', './b.js', 'd', './a.json', './a.handlebars']);
     pkg.files['b.js'].dependencies.should.eql(['b', 'c']);
-    pkgDeps['b'].should.eql(pkg.get('b@1.1.0'));
-    pkgDeps['c'].should.eql(pkg.get('c@1.1.1'));
-    pkgDeps['d'].should.eql(pkg.get('d@0.1.1'));
-    pkgDevDeps['e'].should.eql(pkg.get('e@1.1.0'));
+    pkgDeps['b'].should.eql(pkg.getPackage('b@1.1.0'));
+    pkgDeps['c'].should.eql(pkg.getPackage('c@1.1.1'));
+    pkgDeps['d'].should.eql(pkg.getPackage('d@0.1.1'));
+    pkgDevDeps['e'].should.eql(pkg.getPackage('e@1.1.0'));
 
-    var b = pkg.get('b@1.1.0');
-    b.fileCache.should.equal(pkg.fileCache);
+    var b = pkg.getPackage('b@1.1.0');
+    b._files.should.equal(pkg._files);
     var bDeps = b.dependencies;
     b.main.should.eql('src/b.js');
     b.name.should.eql('b');
     b.version.should.eql('1.1.0');
     b.files['src/b.js'].dependencies.should.eql(['c', './b.tpl']);
-    bDeps['c'].should.eql(pkg.get('c@1.1.1'));
-    bDeps['d'].should.eql(pkg.get('d@0.1.0'));
+    bDeps['c'].should.eql(pkg.getPackage('c@1.1.1'));
+    bDeps['d'].should.eql(pkg.getPackage('d@0.1.0'));
 
-    var c = pkg.get('c@1.1.1');
-    c.fileCache.should.equal(pkg.fileCache);
+    var c = pkg.getPackage('c@1.1.1');
+    c._files.should.equal(pkg._files);
     c.main.should.eql('index.js');
     c.name.should.eql('c');
     c.version.should.eql('1.1.1');
     c.files['index.js'].dependencies.should.eql(['d']);
 
-    var d1 = pkg.get('d@0.1.0');
-    d1.fileCache.should.equal(pkg.fileCache);
+    var d1 = pkg.getPackage('d@0.1.0');
+    d1._files.should.equal(pkg._files);
     d1.main.should.eql('index.js');
     d1.name.should.eql('d');
     d1.version.should.eql('0.1.0');
     d1.files['index.js'].dependencies.should.eql([]);
 
-    var d2 = pkg.get('d@0.1.1');
-    d2.fileCache.should.equal(pkg.fileCache);
+    var d2 = pkg.getPackage('d@0.1.1');
+    d2._files.should.equal(pkg._files);
     d2.main.should.eql('index.js');
     d2.name.should.eql('d');
     d2.version.should.eql('0.1.1');
     d2.files['index.js'].dependencies.should.eql([]);
 
-    var e = pkg.get('e@1.1.0');
-    e.fileCache.should.equal(pkg.fileCache);
+    var e = pkg.getPackage('e@1.1.0');
+    e._files.should.equal(pkg._files);
     e.main.should.eql('src/e.js');
     e.name.should.eql('e');
     e.version.should.eql('1.1.0');
@@ -72,8 +72,8 @@ describe('Father.SpmPackage', function() {
 
   it('output', function() {
     var pkg = getPackage('output');
-    pkg.output.should.eql(['a.js', 'b.js', 'glob/c.js']);
-    Object.keys(pkg.files).should.eql(['index.js','a.js', 'a1.js', 'b.js', 'b1.js', 'glob/c.js', 'c1.js']);
+    pkg.output.should.eql(['a.js', 'b.js', 'glob/c.js', 'a.jpg']);
+    Object.keys(pkg.files).should.eql(['index.js','a.js', 'a1.js', 'b.js', 'b1.js', 'glob/c.js', 'c1.js', 'a.jpg']);
   });
 
   it('output glob duplicate', function() {
@@ -86,7 +86,7 @@ describe('Father.SpmPackage', function() {
     var pkg = getPackage('other-entry', {entry: ['a.js']});
     Object.keys(pkg.files).should.eql(['index.js', 'a.js', 'b.js']);
 
-    var pkgB = pkg.get('b@1.0.0');
+    var pkgB = pkg.getPackage('b@1.0.0');
     Object.keys(pkgB.files).should.eql(['index.js', 'a.js']);
   });
 
@@ -94,7 +94,7 @@ describe('Father.SpmPackage', function() {
     var pkg = getPackage('other-entry', {entry: ['./glob/*.js']});
     Object.keys(pkg.files).should.eql(['index.js', 'glob/c.js', 'b.js']);
 
-    var pkgB = pkg.get('b@1.0.0');
+    var pkgB = pkg.getPackage('b@1.0.0');
     Object.keys(pkgB.files).should.eql(['index.js', 'a.js']);
   });
 
@@ -142,40 +142,37 @@ describe('Father.SpmPackage', function() {
     var pkg = getPackage('cascade');
     var pkgDeps = pkg.dependencies;
     Object.keys(pkgDeps).should.eql(['b']);
-    pkgDeps['b'].should.eql(pkg.get('b@1.1.0'));
+    pkgDeps['b'].should.eql(pkg.getPackage('b@1.1.0'));
     pkg.files['a.js'].dependencies.should.eql(['b']);
   });
 
-  it('get method', function() {
+  it('getPackage method', function() {
     var pkg = getPackage('normal');
-    var b = pkg.get('b@1.1.0');
+    var b = pkg.getPackage('b@1.1.0');
     b.main.should.eql('src/b.js');
     b.name.should.eql('b');
     b.version.should.eql('1.1.0');
     Object.keys(b.dependencies).should.eql(['d', 'c']);
     Object.keys(b.files).should.eql(['src/b.js', 'src/b.tpl']);
 
-    var self = pkg.get('a@1.0.0');
+    var self = pkg.getPackage('a@1.0.0');
     self.should.equal(pkg);
   });
 
-  xit('set method', function() {
+  xit('setPackage method', function() {
     var pkg = getPackage('normal');
-    pkg.set({id: 'b@1.1.0'});
-    pkg.get('b@1.1.0').should.eql({id: 'b@1.1.0'});
+    pkg.setPackage({id: 'b@1.1.0'});
+    pkg.getPackage('b@1.1.0').should.eql({id: 'b@1.1.0'});
 
-    pkg._parse();
-    pkg.get('b@1.1.0').should.eql({id: 'b@1.1.0'});
-
-    pkg.set({id: 'b@1.1.0', name: 'b'});
-    pkg.get('b@1.1.0').should.eql({id: 'b@1.1.0'});
+    pkg.setPackage({id: 'b@1.1.0', name: 'b'});
+    pkg.getPackage('b@1.1.0').should.eql({id: 'b@1.1.0'});
   });
 
   it('getPackages method', function() {
     var pkg = getPackage('normal');
     var pkgs = pkg.getPackages();
 
-    var b = pkg.get('b@1.1.0');
+    var b = pkg.getPackage('b@1.1.0');
     pkgs['b@1.1.0'].should.equal(b);
 
     Object.keys(b.getPackages()).should.eql([
@@ -186,6 +183,36 @@ describe('Father.SpmPackage', function() {
       'e@1.1.0'
     ]);
   });
+
+  it('getFile method', function() {
+    var filepath = join(base, 'normal/sea-modules/d/0.1.0/index.js');
+    var pkg = getPackage('normal');
+    var file = pkg.getFile(filepath);
+    file.path.should.eql(filepath);
+    file.pkg.id.should.eql('d@0.1.0');
+    (pkg.getFile('notexists.js') === null).should.be.true;
+  });
+
+  it('getFiles method', function() {
+    var pkg = getPackage('normal');
+    var files = pkg.getFiles();
+    Object.keys(files).map(function(item) {
+      return item.replace(join(base, 'normal'), '');
+    }).should.eql([
+      '/sea-modules/d/0.1.0/index.js',
+      '/sea-modules/c/1.1.1/index.js',
+      '/sea-modules/b/1.1.0/src/b.js',
+      '/sea-modules/b/1.1.0/src/b.tpl',
+      '/sea-modules/d/0.1.1/index.js',
+      '/sea-modules/e/1.1.0/src/e.js',
+      '/a.js',
+      '/b.js',
+      '/a.json',
+      '/a.handlebars',
+      '/c.js'
+    ]);
+  });
+
 
   it('require other extension', function() {
     var pkg = getPackage('require-other-ext');
@@ -253,48 +280,67 @@ describe('Father.SpmPackage', function() {
       'no-exist': {name: 'a'}
     };
     should.exists(pkgB.dependencies['import-style']);
-    pkgB.dependencies['import-style'].should.equal(pkg.get('import-style@1.0.0'));
+    pkgB.dependencies['import-style'].should.equal(pkg.getPackage('import-style@1.0.0'));
 
     should.not.exists(pkgB.dependencies['no-exist']);
   });
 
-  it('moduleDir', function() {
-    var pkg = getPackage('module-dir', {moduleDir: 'spm-modules'});
+  it('default moduleDir', function() {
+    var pkg = new SpmPackage(join(base, 'module-dir'));
     pkg.files[pkg.main].dependencies.should.eql(['b']);
   });
 
   it('require file in package', function() {
     var pkg = getPackage('file-in-package');
-    pkg.files['index.js'].dependencies.should.eql([
-      'b',
+    pkg.files['index.js'].lookup(function(finfo) {
+      return finfo.ignore ? false : finfo.pkg.name + '/' + finfo.relative;
+    }).should.eql([
+      'b/index.js',
       'b/a.js',
       'b/lib/index.js',
-      'b/lib/b.js'
+      'b/lib/b.js',
+      'c/c.js',
+      'c/index.js'
     ]);
-    var deps = pkg.files['index.js']._dependencies;
-    Object.keys(deps).length.should.eql(4);
-    deps['b'].pkg.name.should.eql('b');
-    deps['b'].path.should.eql('index.js');
-    deps['b/a'].pkg.name.should.eql('b');
-    deps['b/a'].path.should.eql('a.js');
-    deps['b/lib'].pkg.name.should.eql('b');
-    deps['b/lib'].path.should.eql('lib/index.js');
-    deps['b/lib/b.js'].pkg.name.should.eql('b');
-    deps['b/lib/b.js'].path.should.eql('lib/b.js');
 
-    var pkgB = pkg.get('b@1.0.0');
+    var deps = pkg.files['index.js']._dependencies;
+    Object.keys(deps).length.should.eql(5);
+    deps['b'].pkg.name.should.eql('b');
+    deps['b'].relative.should.eql('index.js');
+    deps['b/a'].pkg.name.should.eql('b');
+    deps['b/a'].relative.should.eql('a.js');
+    deps['b/lib'].pkg.name.should.eql('b');
+    deps['b/lib'].relative.should.eql('lib/index.js');
+    deps['b/lib/b.js'].pkg.name.should.eql('b');
+    deps['b/lib/b.js'].relative.should.eql('lib/b.js');
+
+    var pkgB = pkg.getPackage('b@1.0.0');
     pkgB.files['lib/b.js'].dependencies.should.eql(['c/c.js']);
     should.exist(pkgB.files['a.js']);
     should.exist(pkgB.files['lib/b.js']);
     should.exist(pkgB.files['lib/index.js']);
 
-    var pkgC = pkg.get('c@1.0.0');
+    var pkgC = pkg.getPackage('c@1.0.0');
     should.exist(pkgC.files['c.js']);
+  });
+
+  it('ignore file in package', function() {
+    var pkg = getPackage('file-in-package', {
+      ignore: ['c']
+    });
+    pkg.files['index.js'].lookup(function(finfo) {
+      return finfo.ignore ? false : finfo.pkg.name + '/' + finfo.relative;
+    }).should.eql([
+      'b/index.js',
+      'b/a.js',
+      'b/lib/index.js',
+      'b/lib/b.js'
+    ]);
   });
 
   it('unknown name', function() {
     (function() {
-      getPackage('unknown-name')._parse();
+      getPackage('unknown-name');
     }).should.throw('unknown name /a required by test/fixtures/spm/unknown-name/index.js');
   });
 
@@ -302,31 +348,31 @@ describe('Father.SpmPackage', function() {
 
     it('not found ./b', function() {
       (function() {
-        getPackage('not-found')._parse();
+        getPackage('not-found');
       }).should.throw('test/fixtures/spm/not-found/b not found that required by test/fixtures/spm/not-found/index.js');
     });
 
     it('not found ./b.js', function() {
       (function() {
-        getPackage('not-found2')._parse();
+        getPackage('not-found2');
       }).should.throw('test/fixtures/spm/not-found2/b.js not found that required by test/fixtures/spm/not-found2/index.js');
     });
 
     it('no matched version', function() {
       (function() {
-        getPackage('unmatch-version')._parse();
+        getPackage('unmatch-version');
       }).should.throw('no matched version of a');
     });
 
     it('detect main type', function() {
       (function() {
-        getPackage('main-type-error')._parse();
+        getPackage('main-type-error');
       }).should.throw('pkg.spm.main should be string.');
     });
 
     it('detect output type', function() {
       (function() {
-        getPackage('output-type-error')._parse();
+        getPackage('output-type-error');
       }).should.throw('pkg.spm.output should be array.');
     });
 
@@ -338,18 +384,20 @@ describe('Father.SpmPackage', function() {
 
     it('no installed package but required', function() {
       (function() {
-        getPackage('no-installed-package')._parse();
+        getPackage('no-installed-package');
       }).should.throw('b is not in dependencies but required by test/fixtures/spm/no-installed-package/index.js');
     });
 
     it('recursive', function() {
       (function() {
-        getPackage('recursive')._parse();
+        getPackage('recursive');
       }).should.throw('found test/fixtures/spm/recursive/index.js has recursive dependency');
     });
   });
 });
 
 function getPackage(name, options) {
+  options = options || {};
+  if (!options.moduleDir) options.moduleDir = 'sea-modules';
   return new SpmPackage(join(base, name), options);
 }