+++ HISTORY.md
@@ -2,6 +2,15 @@
 
 ---
 
+## 0.10.0
+
+- use spm_modules as moduleDir by default
+- parse package when initialization, not lazy
+- file object extend vinyl
+- change api: get -> getPackage, set -> setPackage
+- new api: getFile, getFiles
+- file will not be parsed when ignore
+
 ## 0.9.7
 
 Improve require priority